refactor: reassign components and create barrel files
