+++ src/shared/components/Badge/index.ts
@@ -1,0 +1,2 @@
+export * from "./Badge.styles";
+export * from "./Badge";
