+++ src/shared/components/SEOHead/index.ts
@@ -1,0 +1,1 @@
+export * from "./SEOHead";
