+++ src/shared/components/VideoPlayer/index.ts
@@ -1,0 +1,1 @@
+export * from "./VideoPlayer";
